Delete unused file copy in resubmission
Closes #340. As mentioned in the issue, there is some nuance here around whether
it's more useful to perform the copy and allow the user to modify the initial
file or rewrite the file from memory. I'm still not totally decided on that
issue[^1], but I think it's still worth deleting this unused code to make it clear
where the resubmitted files are coming from.

[^1]: I am leaning toward preserving the current behavior, though, in order to
maintain the accuracy of the checkpoints. That's another reason for the simple
approach in this PR. It will be a bigger, separate feature to add full support
for modifying individual file templates at runtime.