Fix unused code (#395)
I tried using `expect(unused)` here, but it didn't work for some reason.
`allow` is fine for now. It looks like these are only used for pretty
printing in a single test, so there's probably a better fix here (moving
them to that test's module or just inlining the printing logic), but
this will work for now.